requirements.txt
fastapi
uvicorn
httpx
pyswisseph
pydantic
python-dotenv
pytest
pytest-asyncio
